+++ ext/gyro/thread.c
@@ -7,7 +7,9 @@ static ID ID_empty;
 static ID ID_fiber_ref_count;
 static ID ID_ivar_event_selector_proc;
 static ID ID_ivar_event_selector;
+static ID ID_ivar_join_wait_queue;
 static ID ID_ivar_main_fiber;
+static ID ID_ivar_terminated;
 static ID ID_pop;
 static ID ID_push;
 static ID ID_run_queue;
@@ -248,15 +250,19 @@ VALUE Thread_fiber_break_out_of_ev_loop(VALUE self, VALUE fiber, VALUE resume_ob
 }
 
 VALUE Thread_join_perform(VALUE self) {
-  if (!RTEST(rb_funcall(self, rb_intern("alive?"), 0))) {
+  if (RTEST(rb_ivar_get(self, ID_ivar_terminated))) {
     return self;
   }
 
   VALUE async = rb_funcall(cGyro_Async, ID_new, 0);
-  VALUE wait_queue = rb_ivar_get(self, rb_intern("@join_wait_queue"));
-
+  // The call to Gyro::Async.new may have caused a thread context switch, which
+  // could possibly lead to a race condition. To prevent this we check again if
+  // the thread has already terminated here.
+  if (RTEST(rb_ivar_get(self, ID_ivar_terminated))) {
+    return self;
+  }
+  VALUE wait_queue = rb_ivar_get(self, ID_ivar_join_wait_queue);
   Gyro_Queue_push(wait_queue, async);
-
   VALUE ret = Gyro_Async_await(async);
   RB_GC_GUARD(async);
   return ret;
@@ -289,7 +295,9 @@ void Init_Thread() {
   ID_fiber_ref_count          = rb_intern("fiber_ref_count");
   ID_ivar_event_selector      = rb_intern("@event_selector");
   ID_ivar_event_selector_proc = rb_intern("@event_selector_proc");
+  ID_ivar_join_wait_queue     = rb_intern("@join_wait_queue");
   ID_ivar_main_fiber          = rb_intern("@main_fiber");
+  ID_ivar_terminated          = rb_intern("@terminated");
   ID_pop                      = rb_intern("pop");
   ID_push                     = rb_intern("push");
   ID_run_queue                = rb_intern("run_queue");